+++ Braintree/Drop-In/BTDropInContentView.m
@@ -35,7 +35,7 @@
                                                                                  metrics:@{@"borderWidth": @(self.theme.borderWidth)}
                                                                                    views:@{@"border": summaryBorderBottom}]];
 
-        self.payPalControl = [[BTPayPalControl alloc] init];
+        self.payPalButton = [[BTPayPalButton alloc] init];
 
         self.cardFormSectionHeader = [[UILabel alloc] init];
 
@@ -52,7 +52,7 @@
         // Add Constraints & Subviews
 
         // Full-Width Views
-        for (UIView *view in @[self.payPalControl, self.selectedPaymentMethodView, self.summaryView, self.ctaControl, self.cardForm]) {
+        for (UIView *view in @[self.payPalButton, self.selectedPaymentMethodView, self.summaryView, self.ctaControl, self.cardForm]) {
             [self addSubview:view];
             view.translatesAutoresizingMaskIntoConstraints = NO;
             [self addConstraints:[NSLayoutConstraint constraintsWithVisualFormat:@"H:|[view]|"
@@ -92,7 +92,7 @@
     NSDictionary *viewBindings = @{
                                    @"activityView": self.activityView,
                                    @"summaryView": self.summaryView,
-                                   @"payPalControl": self.payPalControl,
+                                   @"payPalButton": self.payPalButton,
                                    @"cardFormSectionHeader": self.cardFormSectionHeader,
                                    @"cardForm": self.cardForm,
                                    @"ctaControl": self.ctaControl,
@@ -158,14 +158,14 @@
                     self.activityView.alpha = 0.0f;
                 } completion:^(__unused BOOL finished) {
                     [self setState:newState];
-                    self.payPalControl.alpha = 0.0f;
+                    self.payPalButton.alpha = 0.0f;
                     self.cardForm.alpha = 0.0f;
                     self.cardFormSectionHeader.alpha = 0.0f;
                     self.ctaControl.alpha = 0.0f;
                     [self setNeedsUpdateConstraints];
                     [self layoutIfNeeded];
                     [UIView animateWithDuration:duration animations:^{
-                        self.payPalControl.alpha = 1.0f;
+                        self.payPalButton.alpha = 1.0f;
                         self.cardForm.alpha = 1.0f;
                         self.cardFormSectionHeader.alpha = 1.0f;
                         self.ctaControl.alpha = 1.0f;
@@ -198,9 +198,9 @@
     }
 }
 
-- (void) setHidePayPal:(BOOL)payPalControlHidden{
-    _hidePayPal = payPalControlHidden;
-    self.payPalControl.hidden = payPalControlHidden;
+- (void) setHidePayPal:(BOOL)payPalButtonHidden{
+    _hidePayPal = payPalButtonHidden;
+    self.payPalButton.hidden = payPalButtonHidden;
     [self updateLayout];
 }
 
@@ -209,7 +209,7 @@
     // Reset all to hidden, just for clarity
     self.activityView.hidden = YES;
     self.summaryView.hidden = self.hideSummary;
-    self.payPalControl.hidden = YES;
+    self.payPalButton.hidden = YES;
     self.cardFormSectionHeader.hidden = YES;
     self.cardForm.hidden = YES;
     self.selectedPaymentMethodView.hidden = YES;
@@ -221,7 +221,7 @@
             self.activityView.hidden = YES;
             [self.activityView stopAnimating];
             self.ctaControl.hidden = self.hideCTA;
-            self.payPalControl.hidden = self.hidePayPal ;
+            self.payPalButton.hidden = self.hidePayPal ;
             self.cardFormSectionHeader.hidden = NO;
             self.cardForm.hidden = NO;
             break;
@@ -274,7 +274,7 @@
             return @[[NSString stringWithFormat:@"V:|%@-(35)-[cardFormSectionHeader]-(7)-[cardForm]%@|", summaryViewVisualFormat, ctaControlVisualFormat]];
         } else {
             summaryViewVisualFormat = [NSString stringWithFormat:@"%@-(35)-", summaryViewVisualFormat];
-            return @[[NSString stringWithFormat:@"V:|%@[payPalControl(==40)]-(18)-[cardFormSectionHeader]-(7)-[cardForm]%@|", summaryViewVisualFormat, ctaControlVisualFormat]];
+            return @[[NSString stringWithFormat:@"V:|%@[payPalButton(==40)]-(18)-[cardFormSectionHeader]-(7)-[cardForm]%@|", summaryViewVisualFormat, ctaControlVisualFormat]];
         }
 
     } else {