+++ Braintree/PayPal/Braintree-PayPal-Demo/BraintreePayPalDemoViewController.m
@@ -3,9 +3,9 @@
 
 @interface BraintreePayPalDemoViewController () <UITextFieldDelegate>
 
-@property (weak, nonatomic) IBOutlet BTPayPalControl *payPalControl;
+@property (weak, nonatomic) IBOutlet BTPayPalButton *payPalButton;
 
-@property (nonatomic, strong) IBOutlet NSLayoutConstraint *payPalControlHeightConstraint;
+@property (nonatomic, strong) IBOutlet NSLayoutConstraint *payPalButtonHeightConstraint;
 
 @end
 
@@ -14,42 +14,42 @@
 - (void)viewDidLoad {
     [super viewDidLoad];
     [self.view setTranslatesAutoresizingMaskIntoConstraints:YES];
-    [self.payPalControl setTranslatesAutoresizingMaskIntoConstraints:NO];
+    [self.payPalButton setTranslatesAutoresizingMaskIntoConstraints:NO];
 
-    self.payPalControl.client = [[BTClient alloc] initWithClientToken:[BTClient btPayPal_offlineTestClientToken]];
-    self.payPalControl.presentationDelegate = nil;
-    self.payPalControl.delegate = self;
+    self.payPalButton.client = [[BTClient alloc] initWithClientToken:[BTClient btPayPal_offlineTestClientToken]];
+    self.payPalButton.presentationDelegate = nil;
+    self.payPalButton.delegate = self;
 }
 
 - (IBAction)heightSliderValueDidChange:(UISlider *)slider {
-    self.payPalControlHeightConstraint.constant = slider.value * 100.0f;
+    self.payPalButtonHeightConstraint.constant = slider.value * 100.0f;
     [self.view layoutIfNeeded];
 }
 
 - (IBAction)toggledShouldUseCustomPresentationDelegate:(UISwitch *)sender {
-    self.payPalControl.presentationDelegate = (sender.on ? self : nil);
+    self.payPalButton.presentationDelegate = (sender.on ? self : nil);
 }
 
-# pragma mark - BTPayPalControlViewControllerPresenterDelegate implementation
+# pragma mark - BTPayPalButtonViewControllerPresenterDelegate implementation
 
-- (void)payPalControl:(BTPayPalControl *)control requestsPresentationOfViewController:(UIViewController *)viewController {
+- (void)payPalButton:(BTPayPalButton *)button requestsPresentationOfViewController:(UIViewController *)viewController {
     viewController.modalPresentationStyle = UIModalPresentationFormSheet;
     viewController.modalTransitionStyle = UIModalTransitionStyleFlipHorizontal;
     [self presentViewController:viewController animated:YES completion:nil];
 }
 
-- (void)payPalControl:(BTPayPalControl *)control requestsDismissalOfViewController:(UIViewController *)viewController {
+- (void)payPalButton:(BTPayPalButton *)button requestsDismissalOfViewController:(UIViewController *)viewController {
     [viewController dismissViewControllerAnimated:YES completion:nil];
 }
 
-#pragma mark - BTPayPalControlDelegate implementation
+#pragma mark - BTPayPalButtonDelegate implementation
 
-- (void)payPalControl:(BTPayPalControl *)control didCreatePayPalPaymentMethod:(NSString *)paymentMethod {
-    NSLog(@"payPalControl:%@ addedPaymentMethod:%@", control, paymentMethod);
+- (void)payPalButton:(BTPayPalButton *)button didCreatePayPalPaymentMethod:(NSString *)paymentMethod {
+    NSLog(@"payPalButton:%@ addedPaymentMethod:%@", button, paymentMethod);
 }
 
-- (void)payPalControl:(BTPayPalControl *)control didFailWithError:(NSError *)error {
-    NSLog(@"payPalControl:%@ didFailWithError:%@", control, error);
+- (void)payPalButton:(BTPayPalButton *)button didFailWithError:(NSError *)error {
+    NSLog(@"payPalButton:%@ didFailWithError:%@", button, error);
     [[[UIAlertView alloc] initWithTitle:@"Fail"
                                 message:[error description]
                                delegate:nil