+++ Braintree/PayPal/Braintree-PayPal/Public/Braintree-PayPal.h
@@ -1,5 +1,5 @@
 ///  All-in-one import for the Braintree PayPal SDK
 
 #import <Braintree/BTErrors+BTPayPal.h>
-#import <Braintree/BTPayPalControl.h>
+#import <Braintree/BTPayPalButton.h>
 #import <Braintree/BTPayPalViewController.h>